requirements.txt
amqp==2.6.0
anyjson==0.3.3
asgiref==3.2.7
billiard==3.6.3.0
boto3==1.14.31
botocore==1.17.31
celery==4.4.0
cffi==1.14.0
decouple==0.0.7
dj-database-url==0.5.0
Django==3.0.7
django-autoslug==1.9.7
django-celery==3.3.1
django-celery-results==1.2.1
django-ckeditor==5.9.0
django-crispy-forms==1.9.1
django-heroku==0.3.1
django-inplaceedit==1.4.1
django-js-asset==1.2.2
django-kombu==0.9.4
django-storages==1.9.1
dnspython==1.16.0
docutils==0.15.2
environ==1.0
eventlet==0.25.2
flower==0.9.4
future==0.18.2
gevent==20.6.1
greenlet==0.4.16
gunicorn==20.0.4
humanize==0.5.1
jmespath==0.10.0
kombu==4.6.11
monotonic==1.5
numpy==1.18.5
Pillow==7.1.2
psycopg2==2.8.5
pycparser==2.20
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
redis==3.5.3
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
tornado==6.0.4
urllib3==1.25.10
vine==1.3.0
whitenoise==5.1.0
zope.event==4.4
zope.interface==5.1.0
